refactore view files :q :exit exit q
