+++ SpriteKitIntegration/Views/ContentView+Preview.swift
@@ -1,0 +1,14 @@
+//
+//  ContentView+Preview.swift
+//  SpriteKitIntegration
+//
+//  Created by Juliano on 07/11/23.
+//
+
+import SwiftUI
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView()
+    }
+}
